refactor: MyBookDetail への画面遷移をリファクタ
